+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.7.0] - not released
+### Added
+- Add new anchors module with Sep-24 support
 
 ## [1.6.3] - 2022-09-03
 ### Fixed
@@ -13,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.6.2] - 2022-08-30
 ### Added
 - Add support to Muxed accounts
-- 
+
 ### Fixed
 - Fix path payment component